refactor(cli): extract plugin-contributed command dispatch module
